Use 'uint32' instead of 'int64' for blob sizes everywhere.
Not just in blob.SizedRef, but in blobserver.Fetch and
blobserver.FetchStreaming, too.
Blobs have a max size of 10-32 MB anyway, and the index.Corpus is now using
uint32 to save memory.

+++ pkg/client/enumerate.go
@@ -117,7 +117,7 @@ func (c *Client) EnumerateBlobsOpts(ctx *context.Context, ch chan<- blob.SizedRe
 				return error("item in 'blobs' had invalid blobref.", nil)
 			}
 			select {
-			case ch <- blob.SizedRef{Ref: br, Size: size}:
+			case ch <- blob.SizedRef{Ref: br, Size: uint32(size)}:
 			case <-ctx.Done():
 				return context.ErrCanceled
 			}
